+++ cliboa/scenario/transform/file.py
@@ -774,46 +774,3 @@ class FileArchive(FileBaseTransform):
                 raise InvalidParameter(
                     "'format' must set one of the followings [tar, zip]"
                 )
-
-class ExecuteShellScript(FileBaseTransform):
-    """
-    Hash All Column Value
-    """
-    def __init__(self):
-        super().__init__()
-        self._command = ""
-        self._work_dir = ""
-
-    def command(self, command):
-        self._command = command
-
-    def work_dir(self, work_dir):
-        self._work_dir = work_dir
-
-    def execute(self, *args):
-        # essential parameters check
-        valid = EssentialParameters(
-            self.__class__.__name__,
-            [
-                self._command
-            ],
-        )
-        valid()
-
-        # Setting Up Directory
-        default_dir =os.getcwd()
-
-        if self._work_dir:
-            os.chdir(self._work_dir)
-
-        # Run Commands
-        content = self._command.get("content")
-        file_path = self._command.get("file")
-        if content:
-            for command in content.split("&&"):
-                subprocess.run(command.strip().split(" "))
-        elif file_path:
-            subprocess.call(file_path)
-
-        # Set Directory to default
-        os.chdir(default_dir)
