+++ cliboa/test/__init__.py
@@ -28,6 +28,8 @@ class BaseCliboaTest(TestCase):
         """
         setup resource for test
         """
+        print("AAAAAAAAAAAAAAAAAAAAAAAA")
+        print(os.getcwd())
         os.makedirs("common", exist_ok=True)
         os.makedirs("conf", exist_ok=True)
         os.makedirs("logs", exist_ok=True)
