+++ cliboa/test/scenario/transform/test_file.py
@@ -33,8 +33,7 @@ from cliboa.scenario.transform.file import (
     FileConvert,
     FileDecompress,
     FileDivide,
-    FileRename,
-    ExecuteShellScript
+    FileRename
 )
 from cliboa.test import BaseCliboaTest
 from cliboa.util.exception import (
@@ -988,34 +987,3 @@ class TestFileArchive(TestFileTransform):
         files = glob(os.path.join(result_dir, "foo.zip"))
         assert 1 == len(files)
         assert "foo.zip" == os.path.basename(files[0])
-
-class TestExecuteShellScript(TestFileTransform):
-    def test_inline_script(self):
-        instance = ExecuteShellScript()
-        Helper.set_property(instance, "command", {"content": "touch foo.csv && touch test.csv"})
-        Helper.set_property(instance, "work_dir", self._data_dir)
-        instance.execute()
-
-        files = glob(os.path.join(self._data_dir, "*.csv"))
-        assert 2 == len(files)
-        assert "foo.csv" == os.path.basename(files[0])
-        assert "test.csv" == os.path.basename(files[1])
-
-    def test_file_script(self):
-        test_script_path = os.path.join(self._data_dir, "test.sh")
-        test_file_path = os.path.join(self._data_dir, "foo.csv")
-
-        with open(test_script_path, "w") as t:
-            t.write("#!/bin/sh\n")
-            t.write("\n")
-            t.write(f"touch {test_file_path}")
-
-        subprocess.call(['chmod', '0777', test_script_path])
-
-        instance = ExecuteShellScript()
-        Helper.set_property(instance, "command", {"file": test_script_path})
-        instance.execute()
-
-        files = glob(os.path.join(self._data_dir, "*.csv"))
-        assert 1 == len(files)
-        assert "foo.csv" == os.path.basename(files[0])
